requirements.txt
httpx
PyYAML
extvip
pyopenssl